[clang][contracts] P3850 contract extensions
Squashed net diff of the P3850 work on the contracts-p3850 branch
(1634b387b76f..7cff6bda31a7), applied on top of the efcs implementation and
rebased onto current llvm/llvm-project main.  This is the second of the two
distinct epochs (the P3850 extensions: P3097/P3098/P3099/P3100/P3290/P3400/
P4283/P4298/P4299/P4301, libcontracts, sanitizer routing, and the mirror-test
suite).

Conflict resolutions against upstream drift (on top of the efcs commit):
- Token flags: added KEYCONTRACTSP3400 (0x100000000) and widened the TokenKey
  enum / getKeywordStatus flag type to 64-bit (upstream had consumed the last
  32-bit bit).
- CGExprScalar signed-overflow: kept upstream's new overflow-behavior model
  (getOverflowBehaviorConsideringType / OB_*) and prepended the P3100 implicit
  signed-overflow hooks at each site (EmitMul/EmitAdd/EmitSub/IncDec) + the
  float-cast fpint guard.
- Driver: combined wantsCxxContracts with upstream's addSeparateSectionFlags;
  merged libcontracts into the runtimes lists and the runtimes-build
  auto-enable; UBSan contract routing kept alongside upstream's suppressions.

+++ clang/include/clang/Basic/TokenKinds.h
@@ -18,11 +18,12 @@
 #include "llvm/Support/Compiler.h"
 
 #include <cassert>
+#include <cstdint>
 
 namespace clang {
 
 /// Constants for TokenKinds.def
-enum TokenKey : unsigned {
+enum TokenKey : uint64_t {
 #define EMIT_TOKENKEY
 #include "clang/Basic/BuiltinTraits.inc"
 };
